deployvim idempotency for the zenburn theme installation

+++ scripts/deployvim.sh
@@ -3,6 +3,8 @@
 mkdir -p "$HOME"/.vim/colors
 mkdir -p "$HOME"/.vim/undodir
 
-git clone https://github.com/jnurmine/Zenburn.git "$HOME"/zenburn \
-  && mv -f -v "$HOME"/zenburn/colors/zenburn.vim "$HOME"/.vim/colors/ \
-  && rm -f -r "$HOME"/zenburn
+if [ ! -d "$HOME"/.vim/colors/zenburn.vim ] ; then
+  git clone https://github.com/jnurmine/Zenburn.git "$HOME"/zenburn \
+    && mv -f -v "$HOME"/zenburn/colors/zenburn.vim "$HOME"/.vim/colors/ \
+    && rm -f -r "$HOME"/zenburn
+fi
